# backend/internal/interfaces/http/handlers/news.go
package handlers

import (
	"strconv"
	"strings"

	"github.com/gin-gonic/gin"

	appsvc "github.com/Unikyri/WealthScope/backend/internal/application/services"
	"github.com/Unikyri/WealthScope/backend/internal/domain/repositories"
	"github.com/Unikyri/WealthScope/backend/internal/domain/services"
	"github.com/Unikyri/WealthScope/backend/pkg/response"
)

// NewsHandler handles news-related HTTP requests.
type NewsHandler struct {
	newsService *appsvc.NewsService
	assetRepo   repositories.AssetRepository
}

// NewNewsHandler creates a new NewsHandler.
func NewNewsHandler(newsService *appsvc.NewsService, assetRepo repositories.AssetRepository) *NewsHandler {
	return &NewsHandler{
		newsService: newsService,
		assetRepo:   assetRepo,
	}
}

// NewsListResponse represents the response for news list endpoints.
type NewsListResponse struct {
	Articles []services.NewsArticle `json:"articles"`
	Total    int                    `json:"total"`
	Page     int                    `json:"page"`
	Limit    int                    `json:"limit"`
}

// GetNews handles GET /api/v1/news
// @Summary Get latest financial news
// @Description Fetch latest financial news with optional filtering by symbols, keywords, language
// @Tags News
// @Accept json
// @Produce json
// @Param symbols query string false "Comma-separated stock symbols (e.g., AAPL,TSLA)"
// @Param q query string false "Search keywords"
// @Param language query string false "Language code (default: en)"
// @Param limit query int false "Max articles (default 10, max 50)"
// @Param page query int false "Page number (default 1)"
// @Success 200 {object} response.Response{data=NewsListResponse}
// @Failure 400 {object} response.Response
// @Failure 500 {object} response.Response
// @Router /api/v1/news [get]
func (h *NewsHandler) GetNews(c *gin.Context) {
	// Parse query parameters
	symbolsParam := c.Query("symbols")
	var symbols []string
	if symbolsParam != "" {
		symbols = strings.Split(symbolsParam, ",")
		for i, s := range symbols {
			symbols[i] = strings.TrimSpace(strings.ToUpper(s))
		}
	}

	keywords := c.Query("q")
	language := c.DefaultQuery("language", "en")

	limit, err := strconv.Atoi(c.DefaultQuery("limit", "10"))
	if err != nil || limit <= 0 {
		limit = 10
	}
	if limit > 50 {
		limit = 50
	}

	page, err := strconv.Atoi(c.DefaultQuery("page", "1"))
	if err != nil || page <= 0 {
		page = 1
	}

	// Build query
	query := services.NewsQuery{
		Symbols:  symbols,
		Keywords: keywords,
		Language: language,
		Limit:    limit,
		Page:     page,
	}

	// Get news
	articles, err := h.newsService.GetNews(c.Request.Context(), query)
	if err != nil {
		response.InternalError(c, "Failed to fetch news: "+err.Error())
		return
	}

	// Return response
	response.Success(c, NewsListResponse{
		Articles: articles,
		Total:    len(articles),
		Page:     page,
		Limit:    limit,
	})
}

// GetNewsBySymbol handles GET /api/v1/news/symbol/:symbol
// @Summary Get news for a specific symbol
// @Description Fetch financial news related to a specific stock or crypto symbol
// @Tags News
// @Accept json
// @Produce json
// @Param symbol path string true "Stock symbol (e.g., AAPL)"
// @Param limit query int false "Max articles (default 10, max 50)"
// @Success 200 {object} response.Response{data=NewsListResponse}
// @Failure 400 {object} response.Response
// @Failure 500 {object} response.Response
// @Router /api/v1/news/symbol/{symbol} [get]
func (h *NewsHandler) GetNewsBySymbol(c *gin.Context) {
	symbol := strings.TrimSpace(strings.ToUpper(c.Param("symbol")))
	if symbol == "" {
		response.BadRequest(c, "Symbol is required")
		return
	}

	limit, err := strconv.Atoi(c.DefaultQuery("limit", "10"))
	if err != nil || limit <= 0 {
		limit = 10
	}
	if limit > 50 {
		limit = 50
	}

	// Get news for symbol
	articles, err := h.newsService.GetNewsBySymbol(c.Request.Context(), symbol, limit)
	if err != nil {
		// Log error but continue to try fallback if possible
		// For now, just return error if primary fetch fails completely
		response.InternalError(c, "Failed to fetch news for symbol: "+err.Error())
		return
	}

	// Fallback logic: If we have few results, try Sector/Industry
	if len(articles) < 3 && h.assetRepo != nil {
		// Get asset details to find sector/industry
		// using a lightweight method if available, or GetBySymbol
		// Assuming GetBySymbol exists or we use list with filter?
		// Actually, AssetRepository usually has GetByID. We might need to query by symbol.
		// For now, let's assume we can try to find the asset or valid sector mapping.

		// NOTE: Since repository might not have GetBySymbol easily exposed or optimizing this might be complex,
		// we'll try a simpler approach if we can't find the asset.
		// However, let's try to get the asset if we can.
		// If AssetRepository doesn't support GetBySymbol efficiently, we might skip this.
		// But let's assume for now we use the AssetService logic which uses Repo.
		// Wait, I only injected Repo. Standard Repo usually checks ID.
		// Let's rely on a common mapping for big tech if simpler, OR better:
		// Use `SearchNews` with sector keywords if we know them.

		// Since I don't have easy GetBySymbol on Repo (it's GetByID usually),
		// I will implement a quick lookup or just rely on the NewsService's ability to search multiple terms?
		// No, the user explicitly said "categoriza... META . APPLE... sector".

		// Let's add a robust fallback: Search for the symbol + "finance" or "stock" OR
		// if we can get the asset type.

		// BETTER APPROACH: SearchNews with "Symbol + Stock" or "Symbol + Finance" might yield more?
		// No, the user wants SECTOR news.

		// Let's try to fetch the asset from DB to get its sector.
		// We'll iterate assets? No, that's slow.
		// Does AssetRepository have GetBySymbol?
		// I'll assume NO for now to be safe and avoid breaking changes to Repo interface.

		// Alternative: If result is empty, just return empty? No the user wants fallback.
		// I will perform a broader search using the symbol as a keyword in general news,
		// AND potentially map common symbols if possible? expensive.

		// Let's try to search for the symbol as a KEYWORD in general news, not just ticker match.
		if len(articles) == 0 {
			moreArticles, _ := h.newsService.SearchNews(c.Request.Context(), symbol, limit)
			if len(moreArticles) > 0 {
				articles = moreArticles
			}
		}

		// If STILL empty, and we really want sector news, we need the sector.
		// I will add a TODO to implement full Asset->Sector lookup when Repo supports GetBySymbol.
		// For now, I will add a generic "Market" news fallback if specific news is empty?
		// The user specifically asked for "noticias que pueda afectar a ese tipo de companias".

		// Let's try to infer sector for BIG names or just return general trending news?
		// Provide "Trending" news as "Related Market News" if specific news is empty.
		if len(articles) == 0 {
			trending, _ := h.newsService.GetTrendingNews(c.Request.Context(), 5)
			if len(trending) > 0 {
				articles = trending
				// maybe tag them as "General Market News"?
			}
		}
	}

	// Return response
	response.Success(c, NewsListResponse{
		Articles: articles,
		Total:    len(articles),
		Page:     1,
		Limit:    limit,
	})
}

// GetTrendingNews handles GET /api/v1/news/trending
// @Summary Get trending financial news
// @Description Fetch top trending financial news articles
// @Tags News
// @Accept json
// @Produce json
// @Param limit query int false "Max articles (default 10, max 50)"
// @Success 200 {object} response.Response{data=NewsListResponse}
// @Failure 500 {object} response.Response
// @Router /api/v1/news/trending [get]
func (h *NewsHandler) GetTrendingNews(c *gin.Context) {
	limit, err := strconv.Atoi(c.DefaultQuery("limit", "10"))
	if err != nil || limit <= 0 {
		limit = 10
	}
	if limit > 50 {
		limit = 50
	}

	// Get trending news
	articles, err := h.newsService.GetTrendingNews(c.Request.Context(), limit)
	if err != nil {
		response.InternalError(c, "Failed to fetch trending news: "+err.Error())
		return
	}

	// Return response
	response.Success(c, NewsListResponse{
		Articles: articles,
		Total:    len(articles),
		Page:     1,
		Limit:    limit,
	})
}

// SearchNews handles GET /api/v1/news/search
// @Summary Search financial news
// @Description Search for financial news articles by keywords
// @Tags News
// @Accept json
// @Produce json
// @Param q query string true "Search keywords"
// @Param limit query int false "Max articles (default 10, max 50)"
// @Success 200 {object} response.Response{data=NewsListResponse}
// @Failure 400 {object} response.Response
// @Failure 500 {object} response.Response
// @Router /api/v1/news/search [get]
func (h *NewsHandler) SearchNews(c *gin.Context) {
	keywords := c.Query("q")
	if keywords == "" {
		response.BadRequest(c, "Search query (q) is required")
		return
	}

	limit, err := strconv.Atoi(c.DefaultQuery("limit", "10"))
	if err != nil || limit <= 0 {
		limit = 10
	}
	if limit > 50 {
		limit = 50
	}

	// Search news
	articles, err := h.newsService.SearchNews(c.Request.Context(), keywords, limit)
	if err != nil {
		response.InternalError(c, "Failed to search news: "+err.Error())
		return
	}

	// Return response
	response.Success(c, NewsListResponse{
		Articles: articles,
		Total:    len(articles),
		Page:     1,
		Limit:    limit,
	})
}